Finished
Forgot to commit last night. Committing now before work. Most of the
information about the project can be found in the in-line comments and
in the README. This is a basic ETL + CRUD project for manipulating some
Marvel data. There is very little in the way of data-sanitization
because of time constraints.